+++ src/bin/dual-iir.rs
@@ -352,7 +352,7 @@ mod app {
 
                     // Stream the data.
                     const N: usize = BATCH_SIZE * core::mem::size_of::<i16>()
-                        / core::mem::size_of::<MaybeUninit<u32>>();
+                        / core::mem::size_of::<MaybeUninit<u8>>();
                     generator.add::<_, { N * 4 }>(|buf| {
                         for (data, buf) in adc_samples
                             .iter()
@@ -361,7 +361,7 @@ mod app {
                         {
                             let data = unsafe {
                                 core::slice::from_raw_parts(
-                                    data.as_ptr() as *const MaybeUninit<u32>,
+                                    data.as_ptr() as *const MaybeUninit<u8>,
                                     N,
                                 )
                             };